pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "survey-management-app"
version = "1.0.0"
description = "A comprehensive survey management application with FastAPI backend and React frontend"
readme = "README.md"
requires-python = ">=3.13.7"
license = {text = "MIT"}
authors = [
    {name = "Survey Management Team"}
]
keywords = ["survey", "management", "fastapi", "react", "dynamodb"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Framework :: FastAPI",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "fastapi==0.104.1",
    "uvicorn==0.24.0",
    "boto3==1.35.0",
    "botocore==1.35.0",
    "pydantic==2.5.0",
    "python-dotenv==1.0.0",
    "python-multipart>=0.0.7",
    "fastapi-cors==0.0.2",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "graphene==3.3",
    "graphql-core==3.2.3",
    "starlette-graphene3==0.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "pre-commit>=3.0.0",
]

[project.urls]
Homepage = "https://github.com/shenn45/SurveyManagementReactPythonApp"
Repository = "https://github.com/shenn45/SurveyManagementReactPythonApp"
Issues = "https://github.com/shenn45/SurveyManagementReactPythonApp/issues"

[tool.setuptools.packages.find]
where = ["backend"]
include = ["*"]

[tool.black]
line-length = 88
target-version = ["py313"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
py_version = 313

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
no_implicit_optional = true
strict_optional = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]